requirements.txt
python-nmap
texttable